refactor(plugin-api): move memory-only media extraction into the plugin
extractMediaBlocks and extractMediaBlockMeta have no host consumers — they
move into the plugin, rebuilt on the blessed ContentBlock type and
resolveMediaSourceData (byte-equivalent: lossless base64 round-trip). The
contract keeps only the two shared text projections, each with a host
consumer of its own, and the comment now states their split precisely.
