+++ assistant/src/plugins/defaults/memory/graph/image-ref-utils.ts
@@ -1,8 +1,8 @@
-import { extractMediaBlocks } from "@vellumai/plugin-api";
 import { eq } from "drizzle-orm";
 
 import { getDb } from "../../../../persistence/db-connection.js";
 import { messages } from "../../../../persistence/schema/index.js";
+import { extractMediaBlocks } from "../message-media.js";
 import type { ImageRef } from "./types.js";
 
 /**
